+++ assets/js/theme/Category.js
@@ -6,6 +6,7 @@ import Loading from 'bc-loading';
 import svgIcon from './global/svgIcon';
 import fillFacetRatingStars from './global/fillFacetRatingStars';
 import toggleFacet from './global/toggleFacet';
+import InfiniteScroll from 'infinite-scroll';
 
 export default class Category extends PageManager {
   constructor() {
@@ -22,13 +23,19 @@ export default class Category extends PageManager {
     fillFacetRatingStars();
   }
 
+
   loaded(next) {
     this._initializeFacetedSearch(this.context.listingProductCount);
 
     next();
   }
 
+
+
+
   _bindEvents() {
+    this._infiniteScroll();
+
     this.$body.on('click', '[data-listing-view]', (event) => {
       this._toggleView(event);
     });
@@ -38,6 +45,21 @@ export default class Category extends PageManager {
       $(event.currentTarget).toggleClass('is-open').next().toggleClass('visible');
     });
   }
+
+
+  // infinate scroll for BC powered HTML underlay, only for SEO purposes
+  _infiniteScroll() {
+    const elem = document.querySelector('.listing-wrapper');
+    const infScroll = new InfiniteScroll(elem, {
+        path: '.pagination-link--next',
+        append: '.product-grid-item',
+        history: false,
+    });
+    return infScroll;
+  }
+
+
+
 
   _initializeFacetedSearch(productCount) {
     const loadingOptions = {